Move loading filters from storage logic to a helper
