+++ frontend/src/App.vue
@@ -10,13 +10,11 @@
   import { onBeforeMount, provide } from "vue";
   import { useRoute, useRouter } from "vue-router";
 
-  import { populateCommonStores, dateParser } from "@/etc/helpers";
+  import { populateCommonStores } from "@/etc/helpers";
   import TheHeader from "@/components/UserInterface/TheHeader.vue";
   import authApi from "@/services/api/auth";
   import { useAuthStore } from "@/stores/auth";
-  import { useFilterStore } from "@/stores/filter";
-  import { useAlertTableStore } from "@/stores/alertTable";
-  import { useEventTableStore } from "@/stores/eventTable";
+
 
   import { configuration } from "@/etc/configuration/index";
   import { testConfiguration } from "@/etc/configuration/test/index";
@@ -48,16 +46,6 @@
     // localStorage and populate some of the stores with items from
     // the API that will be used throughout the application.
     if (authStore.isAuthenticated) {
-      const filterStore = useFilterStore();
-      const alertTableStore = useAlertTableStore();
-      const eventTableStore = useEventTableStore();
-      filterStore.$state = JSON.parse(
-        localStorage.getItem("aceFilters"),
-        dateParser,
-      );
-      alertTableStore.stateFiltersLoaded = true;
-      eventTableStore.stateFiltersLoaded = true;
-
       await populateCommonStores();
     }
   });
