+++ frontend/src/etc/helpers.ts
@@ -30,6 +30,9 @@ import { useUserStore } from "@/stores/user";
 import { inputTypes } from "@/etc/constants/base";
 import { isValidDate, isObject } from "@/etc/validators";
 import { nodeThreatRead } from "@/models/nodeThreat";
+import { useFilterStore } from "@/stores/filter";
+import { useAlertTableStore } from "@/stores/alertTable";
+import { useEventTableStore } from "@/stores/eventTable";
 
 export const camelToSnakeCase = (str: string): string =>
   str.replace(/[A-Z]/g, (letter) => `_${letter.toLowerCase()}`);
@@ -125,6 +128,18 @@ export function setUserDefaults(nodeType = "all"): void {
       filterValue: currentUserSettingsStore.queues.alerts,
     });
   }
+}
+
+export function loadFiltersFromStorage(): void {
+  const filterStore = useFilterStore();
+  const alertTableStore = useAlertTableStore();
+  const eventTableStore = useEventTableStore();
+  filterStore.$state = JSON.parse(
+    localStorage.getItem("aceFilters") || "{}",
+    dateParser,
+  );
+  alertTableStore.stateFiltersLoaded = true;
+  eventTableStore.stateFiltersLoaded = true;
 }
 
 // https://stackoverflow.com/a/33928558
